+++ contracts/reputation/migration_test.go
@@ -22,8 +22,6 @@ func TestMigration(t *testing.T) {
 	require.NoError(t, err)
 }
 
-var notaryDisabledKey = []byte("notary")
-
 func testMigrationFromDump(t *testing.T, d *dump.Reader) {
 	// gather values which can't be fetched via contract API
 	var epochs []uint64
@@ -43,25 +41,6 @@ func testMigrationFromDump(t *testing.T, d *dump.Reader) {
 	})
 
 	migration.SkipUnsupportedVersions(t, c)
-
-	v := c.GetStorageItem(notaryDisabledKey)
-	notaryDisabled := len(v) == 1 && v[0] == 1
-
-	readPendingVotes := func() bool {
-		if v := c.GetStorageItem([]byte("ballots")); v != nil {
-			item, err := stackitem.Deserialize(v)
-			require.NoError(t, err)
-			arr, ok := item.Value().([]stackitem.Item)
-			if ok {
-				return len(arr) > 0
-			} else {
-				require.Equal(t, stackitem.Null{}, item)
-			}
-		}
-		return false
-	}
-
-	prevPendingVotes := readPendingVotes()
 
 	// read previous values using contract API
 	readEpochsToTrustValues := func() map[uint64][]stackitem.Item {
@@ -94,20 +73,10 @@ func testMigrationFromDump(t *testing.T, d *dump.Reader) {
 
 	prevEpochsToTrustValues := readEpochsToTrustValues()
 
-	// try to update the contract
-	if notaryDisabled && prevPendingVotes {
-		c.CheckUpdateFail(t, "pending vote detected")
-		return
-	}
-
 	c.CheckUpdateSuccess(t)
 
 	// check that contract was updates as expected
 	newEpochsToTrustValues := readEpochsToTrustValues()
-	newPendingVotes := readPendingVotes()
-
-	require.False(t, newPendingVotes, "there should be no more pending votes")
-	require.Nil(t, c.GetStorageItem(notaryDisabledKey), "notary flag should be removed")
 
 	require.Equal(t, len(prevEpochsToTrustValues), len(newEpochsToTrustValues))
 	for k, vPrev := range prevEpochsToTrustValues {