+++ CHANGELOG.md
@@ -6,10 +6,14 @@ Changelog for NeoFS Contract
 ### Added
 
 ### Changed
+- Minimal required version for upgrade is 0.19.0 now (#500)
 
 ### Updated
 
 ### Removed
+- Obsolete list method from container contract (#500)
+- Unused candidate management methods from neofs contract (#500)
+- Unused InnerRingCandidateFee setting from neofs and netmap contracts (#500)
 
 ### Fixed